Add action to build docs as artifacts
Currently, only docs for lively.lang/array are built. This is because a) in some other modules there are doc-strings with an invalid format, causing the build process of the docs to fail and b) some of the folders inside of the repository contain js files of dependencies and/or excessive @module attributions. This triggers a bug in jsdoc which then causes the JS Heap to fill up, stopping the build process (https://github.com/jsdoc/jsdoc/issues/1368).