collab: migrate comments browser to new components architecture
This migration also changes how the CommentBrowser works in generel. Before, it used to be a Singleton whose creation was also forced in each world, even if the comment functionality was never used.
This allowed to permanently store some state in this object.
Now, no information regarding the comment functionality is stored in worlds or anywhere except the `comments` property of Morphs. Comments are loaded from all Morphs in the world each time the CommentBrowser is opened or its mode (resolved/unresolved comments) is changed.
As of this commit, some of the CommentBrowsers minor functionality is thus broken and some layout bugs exist. They are marked with comments throughout the source code.