collab: fix a bug in the badge counter in the topbar
When deleting a morph with comments, the badge counter would not be updated, since the call to update it was made before we could remove the actual Comment object on the morph. Thus, the counter would not yet be decreased. A parameter now indicates that we update the badge because of deletion and the number of comments is adapted accordingly.