🎨: adapt `Text`/`Label` to work with new renderer
This commit in itself is probably not very expressive. This is work in progress code that was made while developing the new renderer in https://github.com/linusha/lively-next-direct-renderer-dev.
The difference between `Text` and `Label` is conceptually broken onwards. The idea at first was to retire `Label` completely and have a `Text` that can be interactively edited as well as rendered efficiently. While developing the new renderer, this was only implemented half-way, leading to a lot of to be removed hacks and the reimplementation of a (smaller) `Label` class.