🧰: hide scrollbar of properties panel
Depending on the height of ones screeen, the properties panel is too long to show all controls on the screen and thus is scrollable. On some systems, the inclusion of the scrollbar lead to the remaining available width not being what the layouts measured and thus the layouting of some controls were messed up.