🎨: set `lineHeight` to 1 for `textAttributes` of `Icon`
For unclear reasons, we observed that on some systems the Icons (e.g., in the system browser) were offset, while on others, they were laid out correctly. It is possible, that this has something to do with linux/mac or with concrete displays and their resolution/scaling.
However, this mitigates the problem for the most part.