🧰: fix inconsistent state of control buttons in meta text of browser
We ensure the correct visibility of the controls in `setPath()`. Unfortunately, the application of a different master wrecks all in the meantime applied styling, as the application of a master is currently overloaded between changing (i.e., resetting) a master and applying a master while respecting overwritten props. Thus, we need to call `setPath()` after each master application.