🧰: stop fill color picker from closing other open color picker
To make sure that we reopen the normal halo after closing the fill halo (which might replace the halo with the gradient halo) we call `showHaloFor()` upon closing the fill color picker. However, as that is also the mechanism that causes the side bar to focus on a morph, this triggered a closing of all currently open popups. As that is no necessary in the case where we just close the fill color picker, we can disable the connection that causes this here.