👤🧰🎨🗩❄️🗒️📦📂: extract storage/access of/to user data into own module again
Previously, this was implemented via getters on the world. This had the downside, that in some context the access via $world was not possible, i.e., on the dashboard.
Extracting this into an own package again should also allow us to refactor the login-flow with GitHub later on and maybe add more advanced user-related functionality.