📂: add `removeUnusedProjectDependencies` method to `Project`
When combined with the components browser, the reconciliation will automatically insert import statements to dependencies and will probably also trigger an addition of the projects dependencies array in the projects `package.json`. Since one might remove an imported component again, this method will take care of removing all introduced dependencies in the `package.json` which are no longer necessary.