👔🧰🛠️: fix test running in CI
The `_uiInitialized` flag was introduced to fix a problem that could occur in the renderer when one ran the `test.sh` script with `headless=false` and had the dev tools opened in the `puppeteer` `chromium` window in mode where they were not separated from the browser window, i.e., docked to a side of the window. This would somehow crash the renderer, since the opening of the test runner happened too early.
We now await the correct initialization of all UI elements before we start to run tests.