📂🛠️🧰: populate `packageRegistry` with projects from `local_projects`
This commit makes it so that we add the contents of the `local_projects` folder to the `packageRegistry`. This way, we can rely on `flatn` for the resolution of project imports. To import a module from inside of a project, just import from the path beginning with `projectRepoOwner-projectName`.
Since `flatn` relies on the contents of the `package.json` for the name of packages, the `name` key in the `package.json` now needs to contain the owner of a project followed by a dash, followed by the name of the project.
At runtime, we patch that to be only the name of the project again, ensuring that the owner and the dash will be inserted again when saving the `package.json` file via the `saveConfigData` method of `Project`.
This also allows us to bundle projects that depend on other projects.