📂🐚: reset local `package.json` upon opening and auto-add missing deps
We observed that one quite easily ended up in situations where opening a project created conflicts (esp. in `package.json`), like so:

1. Bind the local project against a different lively version than upstream, do not commit the resulting change in `package.json`.
2. The upstream repository needs to be one commit ahead and be bound against a different lively version as well.

Then, opening  a project would lead to a `git pull`, followed by `git stash pop`, which would just throw an error afterwards, since the `stash pop` results in conflicts, but the conflicts would persist anyways. This will then lead to the lively server not starting anymore, since `flatn` dies when stumbling upon the now defect `package.json`.

To mitigate this problem, that is especially crucial when affecting the `package.json`, we now do the following:
If one has uncommited changes in the local `package.json`, we just reset them upon opening a project. Since the other crucial operation one could have done that affects the `package.json` is adding new dependencies, we automatically add all missing dependencies to the project for which import statements are present in the code of the project and for which the dependencies projects can be found locally.