📂🧰: improve separator between owner and name of project on disk
Previously, we just used a single dash to separate between the owner of a repository in which a project resides and the name of the project.
However, since the owner is a GitHub name, this would break when one had a dash in ones name, which is allowed by GitHub. Since double dashes are not allowed, we use that now.