📂: fix race condition in project creation
Previously, when creating a new project, a race condition occurred so that one could end up still waiting for the `VersionChecker` to retrieve the current lively version when writing the config file to disk. In those cases, there would be no bound lively version in the `package.json`.
As asynchronous constructors are not (really) possible, I pulled out the offending code into the `ProjectCreationPrompt`.