🎨: preload font before setting `fontFamily` property
This is an optimization that ensures that a font has been loaded before we set the `fontFamily` property of a morph to a value that requires a font file for correct rendering.
Otherwise, the renderer kicked in multiple times while switching fonts which lead to huge delays of multiple seconds until the correct appearance was visible.