🎨: properly apply inverse transform to text measure node
The `TextLayout` that we use to determine clickPos -> char, character extent and others always assumes to be in an untransformed state. When clicking on a text morph, the transformation is "reversed" inside of the morph, and the untransformed position is then relayed to the document.

Therefore, in the case that we measure inside of the font metric, we need to never take tranforms into account, as they will never exist there.
We need to take them into account when measuring inside an actual `TextLayer`, however. Since The textlayer inherits the transform from its owning morph, we need to inverse the transform for the measuring process, and reverse it afterwards.