❄️: do not fail bundling a project without any assets
Previously, the bundler would fail here remotely, as it tried scanning the contents of the assets directory. Locally, we create this directory, but empty folders are not gittable.
One solution would have been to always create an empty and hidden file inside of the assets folder, to ensure that it also exists when clean cloning the project in CI.
The solution I implemented now just skips the step when the asset folder is non-existent. I like this solution, as it reduces clutter in projects without any assets.
Since the addition of assets will be done through the asset manager in the future, which can then handle the creation of the folder if necessary, I think this is a save solution.