📂🧰🐚: do not create empty commits in projects
Previously, it was possible to create effectively empty commits, since one could always press the save button, which in turn would lead to an increase of the patch number of the projects version, which then could be commited.
We prevent this now, by just not opening the save dialogue when no files in the project have changed.