🎨: prevent cursor from moving text when moving over wrapped line edge
When one moved the cursor with the arrow keys over the edge of a wrapped line, the text would be slightly shifted. The underlying cause for this is that the scroll value of a morph can contain invalid values in so far that the assigned values of e.g.,`scrollLeft` and `scroll.x` are not equal, as `scrollLeft` coerces valid, i.e., actual scrollable values. We use this to prevent transforms that lead to undesirable results previously.