🎨: promote vertices to style props
Previously the component definitions would drop vertices and their reconciliation, since they were no longer considered style props. An easy fix to that is to
promote vertices again to be style props. This however requires us to proivde a reasonable default value for vertices, previously we would just throw an error if none was provided.