🫓: prevent cjs resolver from overriding esm resolution
When importing via require, we invoke the custom  commonjs resolver. The cjs resolver is however also invoked if we import via import() or import ‘…’, which is however just via the originalResolve. Since node does not provide info from which context the resolver is invoked we need to find out on our own in order to prevent the esm and cjs resolvers from interfering with each other.