🫓🧑‍🏫❄️📦📂🛠️: `lively_flatn_env` -> `lively_env`
As we need to set `PUPPETEER_CACHE_DIR` in order to have newer versions of `puppeteer` not rely on arbitrary default directories, the env setup is not solely `flatn` specific anymore.
As changing the name of the bashscript and function breaks already generated builds for projects etc, we keep the old version around to deprecate it sometime later, once we have a system around that enables users to react properly to this change.