🎨: fix backwards selection of wrapped lines being rendered with +1 char
The problem occurred due the fact that we too eagerly classified stuff as multi-line, which caused the selection to grow once to the right, as that is added for multi-line-stuff to cover the line break properly.
We classified all selections in wrapped lines as being multi-line, despite the important part being that the check whether the selection itself also covers multiple lines.