🎨📇: embed skipped prop symbol in serialized expression of polic
When serializing a policy we previously would encode the “overridden” properties by writing their values inside the specs within the generated expression of that policy. This had the problem of generating a policy (which supposedly is a replica of the policy) that would not behave in the exact same manner. A skipped property is different to enforcing a particular value at the lowest level, even if it coincides with the currently skipped value of a prop. This lead to weird reset behavior if a policy was consequtively serialized.