🎨: relax overriding condition for overridden masters
Previously if we overrode a master of a morph in a component definition, it would recursively also override the masters of the submorphs of this morph. This makes sense, since that allows us to more easily replace the styling of entire morph hierarchies. The problem was that this recursive overriding was so strong, that it also overwrote local adjustments of submorphs that we part of the same component definition. So the overridden master of a parent always “won” which is not always what we want. Think of swapping the theme of a dark prompt to be light but also want one of the buttons in the prompt to have a custom master outside of the ones found in the dark or light prompt.