🎨: ignore lastExtent when comparing constraint layouts
Factoring in the lastExtent caused issues during style application: the lastExtent would not match, since the morph was resized, yet the layout was not overridden causing the layout to be reset to the original configuration wich would lead to bogus layout glitches. In turn overriding layouts in response to lastExtent changes is rarely desired since the lastExtent changes at an instant the moment the layout is actually used.