🎨: fix yoga node resetting and adjust reset condition
For not entirely clear reasons, yoga nodes need to be reset more often than expected. Not only do we need to reset yoga in case submorphs are removed from the world, but also we need to sometimes reset them, when the layout axis changes. Else the yoga layout will return inconclusive results.