🎨: rewrite application of yoga layout in response to layout requests
This replaces the previous integration of yoga-layout to support the tiling layouts in lively.next.
Overall the initial approach was to tightly couple morphs and their bounds to the yoga nodes and utilize the yoga nodes like “abstract twins“ to compute the layout. This was problematic, since yoga does not like to be evaluated/executed on a per morph level, but instead works best when the layout is continously updated form the top most layout (the morph on top of a layout composition). As a result, this change tries to adhere more to the execution scheme that yoga prefers, making a bunch of hacky adjustments that were previously found in the code unnessecary.