🎨: avoid check for text style classes if not nessecary
Previously especially in text morphs with a large document such as the editors inside the system browser, overly eager checks for certain text style classes would cause the performance to tank. This can be avoided by first checking wether or not the text morph is hugging its text contents or not. In case of hugging, the measurment via canvas has to be dispatched to each line, since textwrapping is not supported on canvas.