🎨: remove hack to post correct incorrect breakpoint styles
This was a hack that fixed an isse due to incorrectly initialized style properties in the fully synthesized buildspecs. The incorrect initialization would happen due to breakpoints only getting correctly triggered after layout application, however since the buildspec synthesization time does not yet allow layouts to be computed, default values needed to be assumed, causing incorrect breakpoint styling as a result. To correct this, we needed a entry point to correct such cases, to which the addMorphAt() lend itself. However this caused other issues, making this solution untenable.